# docs/swagger.yaml
basePath: /
definitions:
...
  model.HttpError:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  model.HttpSuccess:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.PageResult-model_PluginResp:
    properties:
      records:
        items:
          $ref: '#/definitions/model.PluginResp'
        type: array
      total:
        type: integer
    type: object
...
  model.PluginHeartbeatReq:
    properties:
      name:
        type: string
      status:
        type: string
    required:
    - name
    type: object
  model.PluginRegisterReq:
    properties:
      grpc_address:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      start_time:
        type: string
      status:
        type: string
      version:
        type: string
    required:
    - grpc_address
    - name
    - status
    - version
    type: object
  model.PluginResp:
    properties:
      auto_start:
        type: boolean
      bin_path:
        type: string
      config:
        type: string
      created_at:
        type: string
      dependencies:
        items:
          type: string
        type: array
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      key:
        type: string
      last_error:
        type: string
      last_started_at:
        type: string
      last_stopped_at:
        type: string
      magic_cookie_key:
        type: string
      magic_cookie_value:
        type: string
      name:
        type: string
      protocol_version:
        type: string
      status:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
...
host: localhost:13000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /api/v1/plugin/create:
    post:
      consumes:
      - multipart/form-data
      description: 创建一个新的插件
      parameters:
      - description: 插件文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PluginResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建插件
      tags:
      - 插件
  /api/v1/plugin/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定插件
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除插件
      tags:
      - 插件
  /api/v1/plugin/heartbeat:
    post:
      consumes:
      - application/json
      description: 更新插件的心跳时间
      parameters:
      - description: 插件心跳信息
        in: body
        name: heartbeatInfo
        required: true
        schema:
          $ref: '#/definitions/model.PluginHeartbeatReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 插件心跳
      tags:
      - 插件
  /api/v1/plugin/list:
    get:
      consumes:
      - application/json
      description: 获取所有插件的分页列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-model_PluginResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取插件列表
      tags:
      - 插件
  /api/v1/plugin/query/{id}:
    get:
      consumes:
      - application/json
      description: 查询指定ID的插件详情
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PluginResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询插件
      tags:
      - 插件
  /api/v1/plugin/register:
    post:
      consumes:
      - application/json
      description: 注册新插件到系统
      parameters:
      - description: 插件注册信息
        in: body
        name: pluginInfo
        required: true
        schema:
          $ref: '#/definitions/model.PluginRegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 注册插件
      tags:
      - 插件
  /api/v1/plugin/restart/{id}:
    post:
      consumes:
      - application/json
      description: 重启指定插件
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 重启插件
      tags:
      - 插件
  /api/v1/plugin/start/{id}:
    post:
      consumes:
      - application/json
      description: 启动指定插件
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 启动插件
      tags:
      - 插件
  /api/v1/plugin/stop/{id}:
    post:
      consumes:
      - application/json
      description: 停止指定插件
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 停止插件
      tags:
      - 插件
...
swagger: "2.0"